\name{ribo.prof}
\docType{data}
\alias{ribo.prof}
\title{Sample ribosome profiling data}
\description{
  Sample ribosome profiling data for 1000 genes and 4 samples.
}
\usage{ribo.prof}
\format{A list containing two data frames, test.rna that has sample mRNA
  data and test.rp that has sample RPF data.}
\keyword{datasets}
